+++ addons/main/display3DEN/display3DEN.hpp
@@ -31,17 +31,45 @@ class Display3DEN
                 {
                     w = QUOTE(19 * GRID_W);
                 };
-                class Version: ctrlButton
+                // Manually reconfig Version into a button
+                class Version: ValueX
                 {
                     idc = -1;
+                    type = CT_BUTTON;
                     text = "V";
+                    font = "PuristaLight";
+
                     x = QUOTE(ORIGIN_X_STATUSBAR - 15 * GRID_W - SPACE_X);
                     y = QUOTE(2 * pixelH);
                     w = QUOTE(4 * GRID_W - 4 * pixelW);
                     h = QUOTE(4 * GRID_H - 4 * pixelH);
+                    sizeEx = QUOTE(4.32 * SIZEEX_BASE);
+                    offsetX = 0;
+                    offsetY = 0;
+                    offsetPressedX = "pixelW";
+                    offsetPressedY = "pixelH";
+
                     colorBackground[] = {0, 0, 0, 0.5};
+                    colorBackgroundDisabled[] = {0, 0, 0, 0.5};
+                    colorBackgroundActive[] = {COLOR_ACTIVE_RGB, 1};
+                    colorBorder[] = {0, 0, 0, 0};
+                    colorDisabled[] = {1, 1, 1, 0.25};
+                    colorFocused[] = {COLOR_ACTIVE_RGB, 1};
+                    colorShadow[] = {0, 0, 0, 0};
+
                     onLoad = "_this # 0 ctrlSetToolTip format ['%1.%2', (productVersion select 2) * 0.01 toFixed 2, productVersion select 3]";
                     onButtonClick = "call ENH_fnc_productInfo";
+                    soundClick[] = {"\A3\ui_f\data\sound\RscButton\soundClick", 0.09, 1};
+                    soundEnter[] = {"\A3\ui_f\data\sound\RscButton\soundEnter", 0.09, 1};
+                    soundEscape[] = {"\A3\ui_f\data\sound\RscButton\soundEscape", 0.09, 1};
+                    soundPush[] = {"\A3\ui_f\data\sound\RscButton\soundPush", 0.09, 1};
+
+                    borderSize = 0;
+                    period = 0;
+                    periodFocus = 2;
+                    periodOver = 0.5;
+                    style = "0x02 + 0xC0";
+                    shadow = 1;
                 };
                 class TextY: TextX
                 {